feature: first draft of `Promiser` and `BulkPromiser` components
The `Promiser` should be capable of managing a simple promise returning function, e.g., an api fetch call. Any passed children to that promise will be able to render a loading, error and success state. For more complex fetching the `BulkPromiser` is created. It it capable of handling multiple promises at once using the same structure as `Promiser`.